# docs/swagger.yaml
definitions:
  model.Song:
    properties:
      group:
        type: string
      id:
        type: integer
      link:
        type: string
      releaseDate:
        type: string
      song:
        type: string
      text:
        type: string
    type: object
info:
  contact: { }
  version: ""
  title: ""
paths:
  /songs:
    get:
      description: Get songs with filtering and pagination
      parameters:
      - description: Filter by group
        in: query
        name: group
        type: string
      - description: Filter by song name
        in: query
        name: song
        type: string
      - description: Filter by release date
        in: query
        name: releaseDate
        type: string
      - description: Filter by link
        in: query
        name: link
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 10)
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Song'
            type: array
      summary: Get songs
      tags:
      - songs
    post:
      description: Add a new song with details from external API
      parameters:
      - description: Song data
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/model.Song'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Song'
      summary: Create a song
      tags:
      - songs
  /songs/{id}:
    delete:
      description: Delete a song by ID
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Delete a song
      tags:
      - songs
    put:
      description: Update song details
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated song data
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/model.Song'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Song'
      summary: Update a song
      tags:
      - songs
  /songs/{id}/text:
    get:
      description: Get song text paginated by verses
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Verses per page (default 5)
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      summary: Get song text
      tags:
      - songs
swagger: "2.0"